Cast MYSQLPORT to integer for database connection

+++ display_students.php
@@ -4,7 +4,7 @@ $conn = new mysqli(
     getenv('MYSQLUSER'),
     getenv('MYSQLPASSWORD'),
     getenv('MYSQLDATABASE'),
-    getenv('MYSQLPORT')
+    (int) getenv("MYSQLPORT")
 );
 if($conn->connect_error){
     die("DB Error");
